requirements.txt
pandas>=1.5.0,<2.1.0
numpy>=1.21.0,<2.0.0
tensorflow>=2.10.0,<=2.13.0
scikit-learn>=1.0.0,<1.4.0
python-dotenv>=0.19.0,<2.0.0
fredapi>=0.4.3,<1.0.0
pandas-market-calendars>=2.1,<5.0.0
matplotlib>=3.7.0,<4.0.0
scipy>=1.9.0,<1.11.0
statsmodels>=0.14.0,<1.0.0
requests>=2.28.0,<3.0.0
azure-storage-blob>=12.14.0,<13.0.0